backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.45
psycopg[binary]==3.3.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.2
email-validator==2.2.0
beautifulsoup4==4.12.3
httpx==0.27.2
curl-cffi==0.10.0
playwright==1.50.0
jinja2==3.1.4
apscheduler==3.10.4
